Update logic for interval in minutes to handle both ints and doubles, update learn more layout design

+++ insidead/src/main/java/com/streann/insidead/utils/CampaignsFilterUtil.kt
@@ -48,7 +48,7 @@ object CampaignsFilterUtil {
 
         val intervalInMinutes =
             activeCampaign?.properties?.get("intervalInMinutes")
-        val intervalInMillis = intervalInMinutes?.toLong()?.let {
+        val intervalInMillis = intervalInMinutes?.toFloat()?.let {
             Helper.getMillisFromMinutes(it)
         }
         InsideAdSdk.intervalInMinutes = intervalInMillis ?: 0
